// src/core/store/actions/constans.d.ts
export enum ActionTypes {
  SIGN_IN = 'SIGN_IN',
  SIGN_UP = 'SIGN_UP',
  SIGN_OUT = 'SIGN_OUT',
  RESET_PASSW = 'RESET_PASSW',
  RESET_PASSW = 'RESET_PASSW',
  SIGN_ERROR = 'SIGN_ERROR',
  ASYNC_SIGN_IN = 'ASYNC_SIGN_IN',
  ASYNC_SIGN_OUT = 'ASYNC_SIGN_OUT',
  ASYNC_SIGN_IN_GOOGLE = 'ASYNC_SIGN_IN_GOOGLE',
  ASYNC_SIGN_UP = 'ASYNC_SIGN_UP',
  ASYNC_RESET_PASSWORD = 'ASYNC_RESET_PASSWORD',
  GET_CURRENCY_LIST = 'GET_CURRENCY_LIST',
  ASYNC_GET_CURRENCY_LIST = 'ASYNC_GET_CURRENCY_LIST',
  GET_CURRENCY_COURSE = ' GET_CURRENCY_COURSE',
  ASYNC_GET_CURRENCY_COURSE = 'ASYNC_GET_CURRENCY_COURSE',
  GET_CURRENCY_COURSE_LIST = 'GET_CURRENCY_COURSE_LIST',
  ASYNC_GET_CURRENCY_COURSE_LIST = 'ASYNC_GET_CURRENCY_COURSE_LIST',
  GET_CURRENCY_ERROR = 'GET_CURRENCY_ERROR',
  GET_CURRENCY_LIST_FIREBASE = 'GET_CURRENCY_LIST_FIREBASE',
  ASYNC_GET_CURRENCY_LIST_FIREBASE = 'ASYNC_GET_CURRENCY_LIST_FIREBASE',
}